requirements.txt
Authlib==1.6.4
blinker==1.9.0
boto3==1.40.38
botocore==1.40.38
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
ecdsa==0.19.1
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jose==1.0.0
MarkupSafe==3.0.2
pillow==11.3.0
pyasn1==0.6.1
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
gunicorn